services/auth_service/users/views.py
from rest_framework import generics, permissions

from users.serializers import RegisterSerializer, UserProfileSerializer


class RegisterView(generics.CreateAPIView):
	serializer_class = RegisterSerializer
	permission_classes = [permissions.AllowAny]


class MeView(generics.RetrieveAPIView):
	serializer_class = UserProfileSerializer

	def get_object(self):
		return self.request.user
