services/auth_service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "mysqlclient>=2.2.8",
    "python-dotenv>=1.2.2",
]
